+++ src/controller/modProdController.java
@@ -15,6 +15,7 @@ import model.*;
 
 import java.io.IOException;
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 import static controller.mainFormController.getModPartIndex;
@@ -90,20 +91,25 @@ public class modProdController implements Initializable {
     }
 
     public void onModProdRemoveBtn(ActionEvent actionEvent) {
-        try {
-            Part p = (Part) modProdAscPartTable.getSelectionModel().getSelectedItem();
-            if (ascParts.contains(p)) {
-                Product.deleteAssociatedPart(p);
-                ascParts.remove(p);
-                modProdAscPartTable.setItems(ascParts);
-            }
-        }
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Products");
+        alert.setContentText("Are you sure you want to remove this part?");
 
-        catch (NullPointerException e){
-            Alert alert = new Alert(Alert.AlertType.ERROR);
-            alert.setTitle("Error");
-            alert.setContentText("Please select a part. ");
-            alert.show();
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            try {
+                Part p = (Part) modProdAscPartTable.getSelectionModel().getSelectedItem();
+                if (ascParts.contains(p)) {
+                    Product.deleteAssociatedPart(p);
+                    ascParts.remove(p);
+                    modProdAscPartTable.setItems(ascParts);
+                }
+            } catch (NullPointerException e) {
+                Alert err = new Alert(Alert.AlertType.ERROR);
+                err.setTitle("Error");
+                err.setContentText("Please select a part. ");
+                err.show();
+            }
         }
     }
